+++ tauri/src/main.rs
@@ -1,4 +1,5 @@
 mod app;
+mod components;
 mod game;
 
 use app::*;
